+++ seantis/people/tests/test_list.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 from plone import api
 
 from Products.CMFPlone.interfaces.constrains import IConstrainTypes, ENABLED